#2225 :truck: move common composable files in ergebnismeldung/common
