+++ wls-gui-wahllokalsystem/src/composables/ergebnisermittlung/obwStapelCUtils.ts
@@ -6,7 +6,7 @@ import { computed, ref } from "vue";
 
 import { useLogging } from "@/composables/common/logging.ts";
 import { useMathUtils } from "@/composables/common/mathUtils.ts";
-import { useErgebnisUtils } from "@/composables/ergebnismeldung/ergebnisUtils.ts";
+import { useErgebnisUtils } from "@/composables/ergebnismeldung/common/ergebnisUtils.ts";
 import { useErgebnismeldungStore } from "@/stores/ergebnismeldungStore.ts";
 import { useWahlvorschlaegeStore } from "@/stores/wahlvorschlaegeStore.ts";
 import { StapelArtEnum } from "@/types/ergebnismeldung/StapelArtEnum.ts";
