+++ wls-gui-wahllokalsystem/src/composables/ergebnismeldung/common/aWerteMapper.ts
