+++ wls-gui-wahllokalsystem/src/composables/ergebnismeldung/common/ergebnisAndKandidatUtils.ts
