+++ wls-gui-wahllokalsystem/src/composables/ergebnismeldung/common/ergebnisMapper.ts
