+++ wls-gui-wahllokalsystem/src/composables/ergebnismeldung/common/ergebnisUtils.ts
