+++ wls-gui-wahllokalsystem/src/composables/ergebnismeldung/common/statusMapper.ts
