+++ wls-gui-wahllokalsystem/src/composables/ergebnismeldung/common/wahlscheineMapper.ts
