+++ wls-gui-wahllokalsystem/src/composables/ergebnismeldung/common/wahlscheineService.ts
@@ -6,7 +6,7 @@ import {
 } from "@/api/wls-clients/generated-ergebnismeldung-api";
 import { useCommonApiUtils } from "@/composables/api/commonApiUtils.ts";
 import { useLogging } from "@/composables/common/logging.ts";
-import { useWahlscheineMapper } from "@/composables/ergebnismeldung/wahlscheineMapper.ts";
+import { useWahlscheineMapper } from "@/composables/ergebnismeldung/common/wahlscheineMapper.ts";
 import { useUserNotificationService } from "@/composables/userNotification/userNotificationService.ts";
 import { ERGEBNISMELDUNG_SERVICE_API_URL } from "@/constants.ts";
 import { useWahlenStore } from "@/stores/wahlenStore.ts";
