+++ wls-gui-wahllokalsystem/src/composables/ergebnismeldung/common/wahlvorschlagWithKandidatenErgebnissenMapper.ts
