+++ wls-gui-wahllokalsystem/src/stores/wahlscheineStore.ts
@@ -5,7 +5,7 @@ import { ref } from "vue";
 
 import { useHmrUpdate } from "@/composables/common/hmrUpdate.ts";
 import { useLogging } from "@/composables/common/logging.ts";
-import { useWahlscheineService } from "@/composables/ergebnismeldung/wahlscheineService.ts";
+import { useWahlscheineService } from "@/composables/ergebnismeldung/common/wahlscheineService.ts";
 
 const { registerStoreHMR } = useHmrUpdate();
 const { getWahlscheine, postWahlscheine } = useWahlscheineService();
