+++ wls-gui-wahllokalsystem/tests/composables/tasks/taskFactories/aWerteTaskFactory.spec.ts
@@ -16,7 +16,7 @@ const mockDefinitions = vi.hoisted(() => ({
   getAWerte: vi.fn(),
 }));
 
-vi.mock("@/composables/ergebnismeldung/aWerteService.ts", () => ({
+vi.mock("@/composables/ergebnismeldung/common/aWerteService.ts", () => ({
   useAWerteService: vi.fn().mockImplementation(() => ({
     getAWerte: mockDefinitions.getAWerte,
   })),
