+++ wls-gui-wahllokalsystem/tests/stores/ergebnismeldungStore.spec.ts
@@ -22,7 +22,7 @@ const mockDefinitions = vi.hoisted(() => ({
   postBegruendung: vi.fn(),
 }));
 
-vi.mock("@/composables/ergebnismeldung/ergebnisService.ts", () => ({
+vi.mock("@/composables/ergebnismeldung/common/ergebnisService.ts", () => ({
   useErgebnisService: () => ({
     getErgebnisse: mockDefinitions.getErgebnisse,
     postErgebnisse: mockDefinitions.postErgebnisse,
