+++ wls-gui-wahllokalsystem/tests/views/auszaehlung/mbw/MBWStapelDView.spec.ts
@@ -15,7 +15,7 @@ const mockDefinitions = vi.hoisted(() => ({
   postErgebnisse: vi.fn(),
 }));
 
-vi.mock("@/composables/ergebnismeldung/ergebnisService.ts", () => ({
+vi.mock("@/composables/ergebnismeldung/common/ergebnisService.ts", () => ({
   useErgebnisService: () => ({
     getErgebnisse: mockDefinitions.getErgebnisse,
     postErgebnisse: mockDefinitions.postErgebnisse,
